+++ frontend/app/_layout.tsx
@@ -1,4 +1,8 @@
-import { DarkTheme, DefaultTheme, ThemeProvider } from "@react-navigation/native";
+import {
+  DarkTheme,
+  DefaultTheme,
+  ThemeProvider,
+} from "@react-navigation/native";
 import { useFonts } from "expo-font";
 import { Stack } from "expo-router";
 import { StatusBar } from "expo-status-bar";
@@ -20,22 +24,27 @@ export default function RootLayout() {
   }
 
   return (
-    <ThemeProvider value={colorScheme === "dark" ? CustomDarkTheme : CustomDefaultTheme}>
+    <ThemeProvider
+      value={colorScheme === "dark" ? CustomDarkTheme : CustomDefaultTheme}
+    >
       <Stack screenOptions={{ headerShown: false }}>
         {isSignedIn ? (
           <>
-            {/* <Stack.Screen name="(tabs)" />
-            <Stack.Screen
-              name="ProductPage"
-              options={{ animation: "slide_from_bottom" }}
-            /> */}
+            {/* ADD PRODUCT HIDES TABS */}
+            <Stack screenOptions={{ headerShown: false }}>
+              <Stack.Screen name="(tabs)" />
+              <Stack.Screen
+                name="productPage"
+                options={{ headerShown: false }}
+              />
+            </Stack>
           </>
         ) : (
-          <Stack.Screen name='auth/login' />
+          <Stack.Screen name="auth/login" />
         )}
-        <Stack.Screen name='+not-found' />
+        <Stack.Screen name="+not-found" />
       </Stack>
-      <StatusBar style='light' />
+      <StatusBar style="light" />
     </ThemeProvider>
   );
 }
