+++ frontend/components/home/Banner.tsx
