+++ frontend/components/home/ImageContainer.tsx
